Splitting up the huge file with multiple classes
